refactor: OffersList & OffersByCities to entity Offer
