+++ src/entities/OfferCard/index.tsx
@@ -1,2 +1,4 @@
 export type { OfferCardEntity } from './interfaces';
-export { OfferCard } from './ui';
+export { OfferCard } from './ui/OfferCard';
+export { OffersByCities } from './ui/OffersByCities';
+export { OffersList } from './ui/OffersList';
