+++ src/entities/OfferCard/ui/OfferCard.tsx
