+++ src/entities/Review/index.ts
@@ -1,0 +1,2 @@
+export { Review } from './ui/Review';
+export { ReviewsList } from './ui/ReviewsList';
