--- src/features/OffersList/index.ts
@@ -1,1 +1,0 @@
-export { OffersList } from './ui';
